hipamd: change ihipMalloc flags parameter to amd::MemFlags
The unsigned int parameter required callers to cast every amd::MemFlags
value individually.  Using amd::MemFlags directly eliminates all those
casts and makes the API self-documenting.

Changes:
- hip_internal.hpp: update extern declaration
- hip_memory.cpp: ihipMalloc parameter and body, hipHostMalloc,
  hipExtMallocWithFlags, hipMalloc (Empty sentinel), hipMallocHost
- hip_surface.cpp: drop cast at ihipMalloc call site
- hip_texture.cpp: drop cast at ihipMalloc call site

+++ projects/clr/hipamd/src/hip_surface.cpp
@@ -65,7 +65,7 @@ hipError_t ihipCreateSurfaceObject(hipSurfaceObject_t* pSurfObject,
 
   void* surfObjectBuffer = nullptr;
   hipError_t err =
-      ihipMalloc(&surfObjectBuffer, sizeof(__hip_surface), static_cast<unsigned int>(amd::MemFlags::SvmFineGrain));
+      ihipMalloc(&surfObjectBuffer, sizeof(__hip_surface), amd::MemFlags::SvmFineGrain);
   if (surfObjectBuffer == nullptr || err != hipSuccess) {
     return hipErrorOutOfMemory;
   }
